@@ -1,18 +1,17 @@
-const ns = 'KakaoAuthDemo'
-
 export const AmplifyConfig = {
   Auth: {
     region: 'ap-northeast-2',
-    userPoolId: '',           // TODO: replace with actual userpoolid
-    userPoolWebClientId: '',  // TODO: replace with actual clientid
+    userPoolId: '', // TODO: replace with actual userpoolid
+    userPoolWebClientId: '', // TODO: replace with actual clientid
   },
-}
+};
 
-const ApiHash = '' // TODO: replace with actual api url
-export const ApiUrl = `https://${ApiHash}.execute-api.ap-northeast-2.amazonaws.com`
+const ApiHash = ''; // TODO: replace with actual api url
+export const ApiUrl = `https://${ApiHash}.execute-api.ap-northeast-2.amazonaws.com`;
 
 export const IdentityProvider = {
   Kakao: {
-    AppKey: ``, // TODO: replace with actual kakao client id
-  }
-}
+    AppKey: ``, // TODO: replace with actual kakao Javascript key
+    ApiKey: ``, // TODO: replace with actual kakao REST API key
+  },
+};
